Enum piston_meta::Rule

pub enum Rule {
    Whitespace(Whitespace),
    Tag(Tag),
    UntilAny(UntilAny),
    UntilAnyOrWhitespace(UntilAnyOrWhitespace),
    Text(Text),
    Number(Number),
    Select(Select),
    Sequence(Sequence),
    SeparateBy(Box<SeparateBy>),
    Repeat(Box<Repeat>),
    Lines(Box<Lines>),
    Node(Node),
    Optional(Box<Optional>),
}

A rule describes how some section of a document should be parsed.

Variants

Whitespace

Read whitespace.

Tag

Match against a tag.

UntilAny

Reads until any character.

UntilAnyOrWhitespace

Read until any character or whitespace.

Text

Read text.

Number

Read number.

Select

Select one of the sub rules. If the first one does not succeed, try another and so on. If all sub rules fail, then the rule fails.

Sequence

Run each sub rule in sequence. If any sub rule fails, the rule fails.

SeparateBy

Repeat rule separated by another rule.

Repeat

Repeat rule.

Lines

Repeat rule separated by one or more lines.

Node

Read node.

Optional

Read optional.

Methods

impl Rule

fn parse(&self, tokens: &mut Vec<Range<MetaData>>, state: &TokenizerState, read_token: &ReadToken, refs: &[Rule]) -> ParseResult<TokenizerState>

Parses rule.

fn update_refs(&mut self, names: &[Arc<String>])

Updates replacing names with the references.

The references contains the name, but this can not be borrowed as when the same reference is updated.