Module piston_meta::meta_rules

Building blocks for meta rules.

Structs

Lines
Stores information about lines.

Node
A node reference.

Number
Contains information about number.

Optional
Stores information about optional.

Repeat
Stores inforamtion about separated by.

Select
Stores information about select.

SeparateBy
Stores inforamtion about separated by.

Sequence
Stores information about sequence.

Tag
Stores information about tag.

Text
Stores information about text.

UntilAny
Stores information about reading until whitespace or any of some character.

UntilAnyOrWhitespace
Stores information about reading until whitespace or any of some character.

Whitespace
Stores information about whitespace.

Enums

Rule
A rule describes how some section of a document should be parsed.

Functions

parse
Parses text with rules.

parse_errstr
Parses text with rules, formatting the error as a String.

update_refs
Updates the references such that they point to each other.

Type Definitions

ParseResult
A parse result succeeds with a new state, plus an optional error to replace other errors if it is deeper. The deepest error is likely the most useful.