Crate piston_meta

A DSL parsing library for human readable text documents

Reexports

pub use meta_rules::{parse, parse_errstr};

Modules

bootstrap
Bootstrapping meta rules for changing documents at runtime

json
Write meta data as JSON.

meta_rules
Building blocks for meta rules.

tokenizer
Tracking tokenizer state.

Structs

ParseErrorHandler
Reports error.

Syntax
Stores syntax.

Enums

MetaData
Represents meta data.

ParseError
Errors reporting expected values.

Rule
A rule describes how some section of a document should be parsed.

Functions

load_syntax_data
Convenience method for loading data, using the meta language. Panics if there is an error, and writes error message to standard error output.

stderr_unwrap
When an error happens, reports to standard error and then panics.

syntax
Reads syntax from text.

syntax_errstr
Reads syntax from text, formatting the error as String.

Type Definitions

DebugId
The type of debug id used to track down errors in rules.